Use numpy.row_stack() instead of numpy.append()
In Period.close_candlestick() we were using numpy.append(), but this was
changing the shape of our matrix. numpy.row_stack() does the correct behavior.

This commit also changes Candlestick.close() to return a numpy array, as that
is all that we will be using from that point forward anyway.