Run buy/sell methods as threads
Buy/sell can often take a while to complete. During that period, market
conditions can often change and we may no longer want to buy/sell.

The program was previously blocking until the buy/sell completed so we needed
to move the code to a thread to allow the market to continue to be analyzed
while we're trying to complete the order.