Switch to cryptowat.ch for historical API
GDAX historic data is delayed by several minutes. Until they can sort it
out, we will prefer cryptowat.ch's API for historical data, as it is
more up-to-date. GDAX will still be used if there is an issue with
cryptowatch (they have a rate limit of 2 CPU minutes per hour)